+++ test/stdlib/CastTraps.swift.gyb
@@ -5,7 +5,7 @@
 // FIXME: Casting.cpp has dozens of places to fail a cast. This test does not
 // attempt to enumerate them all.
 
-// XFAIL: linux
+// REQUIRES: objc_interop
 
 import StdlibUnittest
