+++ test/stdlib/FloatingPointIR.swift
@@ -4,9 +4,6 @@
 
 var globalFloat32 : Float32 = 0.0
 var globalFloat64 : Float64 = 0.0
-#if arch(i386) || arch(x86_64)
-var globalFloat80 : Float80 = 0.0
-#endif
 
 @inline(never)
 func acceptFloat32(_ a: Float32) {
@@ -18,28 +15,16 @@ func acceptFloat64(_ a: Float64) {
   globalFloat64 = a
 }
 
-#if arch(i386) || arch(x86_64)
-@inline(never)
-func acceptFloat80(_ a: Float80) {
-  globalFloat80 = a
-}
-#endif
-
 func testConstantFoldFloatLiterals() {
   acceptFloat32(1.0)
   acceptFloat64(1.0)
-#if arch(i386) || arch(x86_64)
-  acceptFloat80(1.0)
-#endif
 }
 
 // i386: call swiftcc void @"$s15FloatingPointIR13acceptFloat32yySfF{{.*}}"(float 1.000000e+00)
 // i386: call swiftcc void @"$s15FloatingPointIR13acceptFloat64yySdF{{.*}}"(double 1.000000e+00)
-// i386: call swiftcc void @"$s15FloatingPointIR13acceptFloat80yys0E0VF{{.*}}"(x86_fp80 0xK3FFF8000000000000000)
 
 // x86_64: call swiftcc void @"$s15FloatingPointIR13acceptFloat32yySfF{{.*}}"(float 1.000000e+00)
 // x86_64: call swiftcc void @"$s15FloatingPointIR13acceptFloat64yySdF{{.*}}"(double 1.000000e+00)
-// x86_64: call swiftcc void @"$s15FloatingPointIR13acceptFloat80yys0E0VF{{.*}}"(x86_fp80 0xK3FFF8000000000000000)
 
 // armv7: call swiftcc void @"$s15FloatingPointIR13acceptFloat32yySfF{{.*}}"(float 1.000000e+00)
 // armv7: call swiftcc void @"$s15FloatingPointIR13acceptFloat64yySdF{{.*}}"(double 1.000000e+00)
