+++ test/stdlib/MathConstants.swift
@@ -4,6 +4,8 @@
 import Darwin
 #elseif os(Linux) || os(FreeBSD) || os(PS4) || os(Android)
 import Glibc
+#elseif os(Windows)
+import MSVCRT
 #endif
 
 _ = M_PI // expected-warning {{is deprecated}}
