+++ test/stdlib/NumericParsing.swift.gyb
@@ -134,7 +134,7 @@ tests.test("FixedWidthInteger/maxUInt64") {
 % for Self in 'Float', 'Double', 'Float80':
 
 % if Self == 'Float80':
-#if arch(i386) || arch(x86_64)
+#if !os(Windows) && (arch(i386) || arch(x86_64))
 % end
 
 tests.test("${Self}/Basics") {
