+++ test/stdlib/Runtime.swift.gyb
@@ -14,6 +14,9 @@ import SwiftShims
 import Darwin
 #elseif os(Linux) || os(FreeBSD) || os(PS4) || os(Android)
 import Glibc
+#elseif os(Windows)
+import MSVCRT
+import WinSDK
 #endif
 
 @_silgen_name("swift_demangle")
@@ -566,14 +569,27 @@ Runtime.test("SwiftError layout constants for LLDB") {
   #else
     _UnimplementedError()
   #endif
+#elseif os(Windows)
+  let hStdlibCore: HMODULE = GetModuleHandleA("swiftCore.dll")!
 #else
   _UnimplementedError()
 #endif
 
+#if os(Windows)
+  let offsetof_SwiftError_typeMetadata: UnsafeRawPointer =
+      unsafeBitCast(GetProcAddress(hStdlibCore,
+                                   "_swift_lldb_offsetof_SwiftError_typeMetadata")!,
+                    to: UnsafeRawPointer.self)
+  let sizeof_SwiftError: UnsafeRawPointer =
+      unsafeBitCast(GetProcAddress(hStdlibCore,
+                                   "_swift_lldb_sizeof_SwiftError")!,
+                    to: UnsafeRawPointer.self)
+#else
   let offsetof_SwiftError_typeMetadata =
     dlsym(RTLD_DEFAULT, "_swift_lldb_offsetof_SwiftError_typeMetadata")!
   let sizeof_SwiftError =
     dlsym(RTLD_DEFAULT, "_swift_lldb_sizeof_SwiftError")!
+#endif
 #if os(macOS) || os(iOS) || os(watchOS) || os(tvOS)
 #if arch(i386) || arch(arm)
   expectEqual(20, offsetof_SwiftError_typeMetadata.load(as: UInt.self))
@@ -582,7 +598,7 @@ Runtime.test("SwiftError layout constants for LLDB") {
   expectEqual(40, offsetof_SwiftError_typeMetadata.load(as: UInt.self))
   expectEqual(72, sizeof_SwiftError.load(as: UInt.self))
 #endif
-#elseif os(Linux) || os(Android)
+#elseif os(Linux) || os(Android) || os(Windows)
   expectEqual(16, offsetof_SwiftError_typeMetadata.load(as: UInt.self))
   expectEqual(32, sizeof_SwiftError.load(as: UInt.self))
 #else
