+++ test/stdlib/simd.swift.gyb
@@ -2,7 +2,7 @@
 // REQUIRES: executable_test
 
 // FIXME: No simd module on linux rdar://problem/20795411
-// XFAIL: linux
+// XFAIL: linux, windows
 
 import simd
 import StdlibUnittest
