+++ test/stdlib/tgmath.swift.gyb
@@ -17,10 +17,12 @@
 // RUN: %line-directive %t/tgmath.swift -- %target-run %t/a.out
 // REQUIRES: executable_test
 
-#if canImport(Darwin)
+#if os(iOS) || os(macOS) || os(tvOS) || os(watchOS)
   import Darwin.C.tgmath
-#else
+#elseif os(Linux) || os(android) || os(FreeBSD) || os(Haiku) || os(PS4)
   import Glibc
+#elseif os(Windows)
+  import MSVCRT
 #endif
 
 #if (arch(i386) || arch(x86_64)) && !os(Windows)
@@ -77,7 +79,9 @@ internal protocol TGMath: BinaryFloatingPoint {
 %end
   static func _remquo(_ x: Self, _ y: Self) -> (Self, Int)
   static func _fma(_ x: Self, _ y: Self, _ z: Self) -> Self
+#if !os(Windows)
   static func _lgamma(_ x: Self) -> (Self, Int)
+#endif
   static func _modf(_ x: Self) -> (Self, Self)
   static func _scalbn(_ x: Self, _ n: Int) -> Self
   static func _frexp(_ x: Self) -> (Self, Int)
@@ -116,8 +120,10 @@ internal extension TGMath {
     expectEqualWithTolerance(0.4041169094348222983238250859191217675, Self._erf(0.375))
     expectEqualWithTolerance(0.5958830905651777016761749140808782324, Self._erfc(0.375))
     expectEqualWithTolerance(2.3704361844166009086464735041766525098, Self._tgamma(0.375))
+#if !os(Windows)
     expectEqualWithTolerance( -0.11775527074107877445136203331798850, Self._lgamma(1.375).0, ulps: 16)
     expectEqual(1, Self._lgamma(1.375).1)
+#endif
     expectEqual(1, Self._ceil(0.375))
     expectEqual(0, Self._floor(0.375))
     expectEqual(0, Self._nearbyint(0.375))
@@ -181,7 +187,9 @@ extension ${T}: TGMath {
 %end
   static func _remquo(_ x: ${T}, _ y: ${T}) -> (${T}, Int) { return remquo(x, y) }
   static func _fma(_ x: ${T}, _ y: ${T}, _ z: ${T}) -> ${T} { return fma(x, y, z) }
+#if !os(Windows)
   static func _lgamma(_ x: ${T}) -> (${T}, Int) { return lgamma(x) }
+#endif
   static func _modf(_ x: ${T}) -> (${T}, ${T}) { return modf(x) }
   static func _scalbn(_ x: ${T}, _ n: Int) -> ${T} { return scalbn(x, n) }
   static func _frexp(_ x: ${T}) -> (${T}, Int) { return frexp(x) }
